pyproject.toml
[tool.poetry]
name = "novikov-staff-bot"
version = "0.1.0"
description = ""
authors = ["Evan Vilagov <evan.vilagov@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
requests = "^2.25.1"
python-dotenv = "^0.17.1"
pyTelegramBotAPI = "^3.7.9"
google-api-python-client = "^2.5.0"
oauth2client = "^4.1.3"
google-auth-oauthlib = "^0.4.4"
google-auth-httplib2 = "^0.1.0"
pygsheets = "^2.0.5"
python-crontab = "^2.5.1"
psycopg2-binary = "^2.8.6"

[tool.poetry.dev-dependencies]
flake8 = "^3.9.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
